requirements.txt
# Core AI framework
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-community==0.3.0
langchain-postgres==0.0.17

# Environment / config
python-dotenv==1.0.0

# Telegram channel
python-telegram-bot==21.0

# Web chat channel (FastAPI + server)
fastapi==0.135.1
uvicorn==0.42.0
pydantic==2.12.5

# Database (Postgres memory)
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.10

# Meta channel (WhatsApp / Instagram / Messenger)
httpx==0.28.1
